+++ test/errors/errors.config.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const config = require('../default.config');
+
+module.exports = {
+  config
+};